+++ src/main/java/dev/gacbl/logicore/core/CreativeTabModule.java
@@ -1,6 +1,7 @@
 package dev.gacbl.logicore.core;
 
 import dev.gacbl.logicore.LogiCore;
+import dev.gacbl.logicore.blocks.battery.basic.BasicBatteryModule;
 import dev.gacbl.logicore.blocks.cloud_interface.CloudInterfaceModule;
 import dev.gacbl.logicore.blocks.compiler.CompilerModule;
 import dev.gacbl.logicore.blocks.computer.ComputerModule;
@@ -41,6 +42,7 @@ public class CreativeTabModule {
                         output.accept(GeneratorModule.GENERATOR.get());
                         output.accept(CloudInterfaceModule.CLOUD_INTERFACE.get());
                         output.accept(WrenchModule.WRENCH.get());
+                        output.accept(BasicBatteryModule.BASIC_BATTERY.get());
                     })
                     .build());